Open-AutoGLM/requirements.txt
Pillow>=12.0.0
openai>=2.9.0
PyYAML>=6.0
numpy>=1.24.0

# For iOS Support
requests>=2.31.0

# For Assertion Support (Optional)
# paddleocr>=2.7.0  # 用于 OCR 文字识别,按需安装

# For Model Deployment

## After installing sglang or vLLM, please run pip install -U transformers again to upgrade to 5.0.0rc0.
## Any dependency conflicts related to Transformers can be ignored.

# sglang>=0.5.6.post1
# vllm>=0.12.0
# transformers>=5.0.0rc0

# Optional: for development
# pytest>=7.0.0
# pre-commit>=4.5.0
# black>=23.0.0
# mypy>=1.0.0
